#31: Finalize Core Action(s)
During the initial development we have created a basic
implementation for the action handling for Somos.

Later on we have refined the API to be more intuitive and
readable, and based on feedback by @mnapoli the handling
for the Action handler and responder was simplified by
making use of PHP-DIs `call()` method.

In order to improve the segregation of responsibilities I have
also moved some of the methods around and introduced an
action invoker message responsible for the actual invoking of
an action and calling of the responder.