#17: Add support for scopes
In order to have a single point where the kernel is instantiated
but still be able to execute command line commands we add
the concept of scopes.

With a scope you can restrict a Message from being handled
depending on the environment where it is called. By default we have the 'cli' and 'web' scope.

With this division we can combine the command runner and
web handling into a single file and as such make it as flexible as
people want.

If people care about performance they can still choose to divide
it into separate files but the default is to be the easiest.