Replaced SimpleBus with Tactician
Because the Tactician library does not require specific interfaces for
their handlers and messages, and has less dependencies, I have chosen to
replace the SimpleBus implementation with that of Tactician.

This allows a stricter API and to have less dependencies; which makes it
easier to integrate Somos in an existing codebase.